+++ board/variscite/mx6var_som/zenux_mcontroller_crc.h
@@ -5,7 +5,7 @@
 
 #ifndef CONFIG_SPL_BUILD
 
-u8 calcBlockCRC(u8* block, u32 blocklen);
+u8 calcBlockCRC(u8* block, u16 blocklen);
 u8 calcByteCRC(u8 crc, u8 data);
 
 #endif /* CONFIG_SPL_BUILD */
